+++ src/backend/routes/userDashboardRoutes.js
@@ -1,0 +1,12 @@
+import express from 'express';
+import { 
+  getUserTicketCounts, 
+  getUserRecentTickets 
+} from '../controllers/userDashboardController.js';
+
+const router = express.Router();
+
+router.get('/counts/:userId', getUserTicketCounts);
+router.get('/recent/:userId', getUserRecentTickets);
+
+export default router;
